Finished

+++ model/binance/kline.go
@@ -1,13 +1,13 @@
 package binance
 
 type Kline struct {
-	OpenTime         uint64
+	OpenTime         float64
 	Open             string
 	High             string
 	Low              string
 	Close            string
 	Volume           string
-	CloseTime        uint64
+	CloseTime        float64
 	QuoteAssetVolume string
-	NumberOfTrades   uint64
+	NumberOfTrades   float64
 }
